backend/source_code/django-api/core/auth/viewsets/refresh.py
from rest_framework.response import Response
from rest_framework_simplejwt.views import TokenRefreshView
from rest_framework_simplejwt.exceptions import TokenError, InvalidToken
from rest_framework import status,viewsets

class RefreshViewSet(TokenRefreshView,viewsets.ViewSet):
    http_method_names=['post']
    def create(self,request,*args,**kwargs):
        serializer = self.get_serializer(data=request.data)
        try:
            serializer.is_valid(raise_exception=True)
        except TokenError as e:
            raise InvalidToken(e.args[0])
        return Response(serializer.validated_data, status=status.HTTP_200_OK)

    