backend/source_code/django-api/requirements.txt
Django==4.0
django-cors-headers==4.2.0
djangorestframework==3.14.0
mysqlclient==2.2.0
psycopg2==2.9.7
psycopg2_binary==2.9.2
